pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stromschlag"
version = "0.1.0"
description = "GUI tool for crafting KDE Plasma and GNOME icon packs"
readme = "README.md"
authors = [
  { name = "bastian" }
]
requires-python = ">=3.11"
dependencies = [
  "PySide6>=6.7",
  "Pillow>=10.3",
  "pyyaml>=6.0"
]

[project.urls]
Homepage = "https://example.com/stromschlag"

[project.scripts]
stromschlag = "stromschlag.app:main"

[tool.setuptools.packages.find]
where = ["src"]
